Moved parallax components into separate folder.
