+++ src/components/Home.js
@@ -5,11 +5,11 @@ import About from './About'
 import Projects from './Projects'
 import Team from './Team'
 import Contact from './Contact'
-import Parallax1 from './Parallax1'
-import Parallax2 from './Parallax2'
-import Parallax3 from './Parallax3'
-import Parallax4 from './Parallax4'
-import Parallax5 from './Parallax5'
+import Parallax1 from './parallax/Parallax1'
+import Parallax2 from './parallax/Parallax2'
+import Parallax3 from './parallax/Parallax3'
+import Parallax4 from './parallax/Parallax4'
+import Parallax5 from './parallax/Parallax5'
 import { Element} from 'react-scroll'
 
 const Home = () => {
